setup game plugin, move events and resources inside
